WIP: this is a broken commit.
The good parts:
not much

The bad parts:
Still looking for a way for child frames which are inside the graph, to send
meesage to the parent. When testing locally, then frames are basically
cross-origin as we run them on different port. Theoretically in prod
they would be behind an authentication proxy and this would be fine.
Its still is a problem, because then when running locally you can not
access window.parent.location.origin.