workelements: separate popover from keyword input
The good is that popover can be dismissed by pressing on the text box.
Its more clear how to summon the popover, this is bugfix for situation
when popover was successor to keyword input and when button background
was active, the popover would still not open, user had to click on text.

The bad is that word index is detached from keyword input, not
highlighting their relation.

The ugly is that elm-bootstrap is still rough, maybe need to rewrite
using just the raw bootstrap classes, but that would also be ugly.