manager: try to fix bad loading of workelements
Workelemnts started to load slowly. When graph is up and makes
request to workelements, the service is not up yet and request
goes to manager breaking UI.

The hack right now is for manager to respond only to "graph" urls

This should have been fixed with depends_on docker-compose setting,
but its not. There is a huge delay in getting initial HTTP response
from workelements.